refactor(catalog): rename installcontrols -> index
